Move html_utils.py to utils folder
